requirements.txt
python-telegram-bot=>12.4.2